+++ releases.md
@@ -1,4 +1,6 @@
 # Release notes
 
 
-**1.0**: initial release implementing LCS and SCS for two vectors of parameterized type.
+**1.1.0**: adds `Pairing` class and `align` method for `SequenceComp` computing a Vector of `Pairing`s.
+
+**1.0.0**: initial release implementing LCS and SCS for two vectors of parameterized type.